refactor: extract generateOrderNumber from lib/db/orders to lib/utils
Move generateOrderNumber() to lib/utils/order-number.ts so the test
no longer imports a file with Cloudflare/D1 dependencies at the
top level. This eliminates the fragile implicit dependency on lazy
evaluation of DB imports during tests.
