+++ lib/db/orders.ts
@@ -4,15 +4,6 @@ import { nanoid } from "nanoid";
 import type { Order, OrderItem } from "@/lib/db/types";
 export type { Order, OrderItem };
 
-export function generateOrderNumber(): string {
-  const chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
-  let result = "ORD-";
-  for (let i = 0; i < 6; i++) {
-    result += chars.charAt(Math.floor(Math.random() * chars.length));
-  }
-  return result;
-}
-
 interface CreateOrderData {
   userId: string;
   orderNumber: string;
